+++ src/bucket.c
@@ -1,5 +1,5 @@
-#include "bucket.h"
-#include "io.h"
+#include "include/bucket.h"
+#include "include/io.h"
 
 /*/////////////*/
 /* Set methods */
