+++ src/include/encode.h
