+++ tests/ps.as
@@ -14,7 +14,7 @@ L1: 	inc L3
 	bne LOOP 
 END: hlt
 .define len = 4
-STR: .string "abcd ef" 
+STR: .string "abcdef" 
 LIST: .data 6, -9, len 
 K: .data 22 
 .extern L3
